requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-docx==1.1.0
pydantic==2.5.0
aiofiles==23.2.1
